research/pyproject.toml
[project]
name = "research"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "linkup-sdk>=0.2.4",
    "loguru>=0.7.3",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.5",
]
